chore: make domain.py simple to read
